added more hooks before searching images, refactoring

+++ zImageOptimizer.sh
@@ -5,41 +5,6 @@
 # License: MIT
 # Version: 0.9.6-dev
 
-# Define default vars
-BINARY_PATHS="/bin /usr/bin /usr/local/bin"
-TMP_PATH="/tmp"
-TOOLS="jpegoptim jpegtran djpeg cjpeg pngcrush optipng pngout advpng gifsicle"
-DEPS_DEBIAN="jpegoptim libjpeg-progs pngcrush optipng advancecomp gifsicle wget autoconf automake libtool make bc"
-DEPS_REDHAT="jpegoptim libjpeg* pngcrush optipng advancecomp gifsicle wget autoconf automake libtool make bc"
-DEPS_MACOS="jpegoptim libjpeg pngcrush optipng advancecomp gifsicle jonof/kenutils/pngout"
-GIT_URL="https://github.com/zevilz/zImageOptimizer"
-TIME_MARKER_PATH=""
-TIME_MARKER_NAME=".timeMarker"
-
-# Min versions of distributions. Must be integer.
-MIN_VERSION_DEBIAN=7
-MIN_VERSION_UBUNTU=14
-MIN_VERSION_FEDORA=24
-MIN_VERSION_RHEL=6
-MIN_VERSION_CENTOS=6
-
-# Min version MacOS (second digit; ex. 10.12.2 == 12).
-MIN_VERSION_MACOS=10
-
-# Spacese separated supported versions of distributions.
-SUPPORTED_VERSIONS_FREEBSD="10.3 10.4 11.1"
-
-if [ "Z$(ps o comm="" -p $(ps o ppid="" -p $$))" == "Zcron" -o \
-     "Z$(ps o comm="" -p $(ps o ppid="" -p $(ps o ppid="" -p $$)))" == "Zcron" ]; then
-	SETCOLOR_SUCCESS=
-	SETCOLOR_FAILURE=
-	SETCOLOR_NORMAL=
-else
-	SETCOLOR_SUCCESS="echo -en \\033[1;32m"
-	SETCOLOR_FAILURE="echo -en \\033[1;31m"
-	SETCOLOR_NORMAL="echo -en \\033[0;39m"
-fi
-
 sayWait()
 {
 	local AMSURE
@@ -214,6 +179,9 @@ installDeps()
 
 	fi
 
+	# Hook: after-check-platform
+	includeExtensions after-check-platform
+
 	if [ $DEBUG -eq 1 ]; then
 		echo "Platform info:"
 		echo
@@ -238,25 +206,66 @@ installDeps()
 		fi
 
 		if [ $PLATFORM == "linux" ]; then
+
+			# Hook: before-install-deps-linux
+			includeExtensions before-install-deps-linux
+
 			if [ $PLATFORM_PKG == "debian" ]; then
+
+				# Hook: before-install-deps-debian
+				includeExtensions before-install-deps-debian
+
 				$SUDO apt-get update
 				$SUDO apt-get install $DEPS_DEBIAN -y
 
+				# Hook: after-install-deps-debian
+				includeExtensions after-install-deps-debian
+
 			elif [ $PLATFORM_PKG == "redhat" ]; then
 
+				# Hook: before-install-deps-redhat
+				includeExtensions before-install-deps-redhat
+
 				if [ $PLATFORM_DISTRIBUTION == "Fedora" ]; then
+
+					# Hook: before-install-deps-redhat-fedora
+					includeExtensions before-install-deps-redhat-fedora
+
 					$SUDO dnf install epel-release -y
 					$SUDO dnf install $DEPS_REDHAT -y
+
+					# Hook: after-install-deps-redhat-fedora
+					includeExtensions after-install-deps-redhat-fedora
+
 				elif [ $PLATFORM_DISTRIBUTION == "RHEL" ]; then
+
+					# Hook: before-install-deps-redhat-rhel
+					includeExtensions before-install-deps-redhat-rhel
+
 					$SUDO yum install https://dl.fedoraproject.org/pub/epel/epel-release-latest-$PLATFORM_VERSION.noarch.rpm -y
 					echo
 					echo -n "Enabling rhel-$PLATFORM_VERSION-server-optional-rpms repository..."
 					$SUDO subscription-manager repos --enable rhel-$PLATFORM_VERSION-server-optional-rpms
 					$SUDO yum install $DEPS_REDHAT -y
+
+					# Hook: after-install-deps-redhat-rhel
+					includeExtensions after-install-deps-redhat-rhel
+
 				else
+
+					# Hook: before-install-deps-redhat-other
+					includeExtensions before-install-deps-redhat-other
+
 					$SUDO yum install epel-release -y
 					$SUDO yum install $DEPS_REDHAT -y
+
+					# Hook: after-install-deps-redhat-other
+					includeExtensions after-install-deps-redhat-other
+
 				fi
+
+				# Hook: after-install-deps-redhat
+				includeExtensions after-install-deps-redhat
 
 				if [[ $PLATFORM_DISTRIBUTION == "CentOS" && $PLATFORM_VERSION -eq 6 || $PLATFORM_DISTRIBUTION == "RHEL" && $PLATFORM_VERSION -eq 6 ]]; then
 					for p in "${!BINARY_PATHS_ARRAY[@]}" ; do
@@ -264,7 +273,7 @@ installDeps()
 							ISSET_pngcrush=1
 						fi
 					done
-					if [ $ISSET_pngcrush -eq 0 ]; then
+					if ! [ -z $ISSET_pngcrush ] && [ $ISSET_pngcrush -eq 0 ]; then
 						wget https://downloads.sourceforge.net/project/pmt/pngcrush/old-versions/1.8/1.8.0/pngcrush-1.8.0.tar.gz
 						tar -zxvf pngcrush-1.8.0.tar.gz
 						rm pngcrush-1.8.0.tar.gz
@@ -280,7 +289,7 @@ installDeps()
 							ISSET_advpng=1
 						fi
 					done
-					if [ $ISSET_advpng -eq 0 ]; then
+					if ! [ -z $ISSET_advpng ] && [ $ISSET_advpng -eq 0 ]; then
 						$SUDO yum install zlib-devel gcc-c++ -y
 						wget https://github.com/amadvance/advancecomp/releases/download/v2.0/advancecomp-2.0.tar.gz
 						tar -zxvf advancecomp-2.0.tar.gz
@@ -322,7 +331,7 @@ installDeps()
 	#			rm -rf mozjpeg
 	#		fi
 
-			if [ $ISSET_pngout -eq 0 ]; then
+			if ! [ -z $ISSET_pngout ] && [ $ISSET_pngout -eq 0 ]; then
 				wget http://static.jonof.id.au/dl/kenutils/pngout-20150319-linux.tar.gz
 				tar -xf pngout-20150319-linux.tar.gz
 				rm pngout-20150319-linux.tar.gz
@@ -334,7 +343,13 @@ installDeps()
 				rm -rf pngout-20150319-linux
 			fi
 
+			# Hook: after-install-deps-linux
+			includeExtensions after-install-deps-linux
+
 		elif [ $PLATFORM == "macos" ]; then
+
+			# Hook: before-install-deps-macos
+			includeExtensions before-install-deps-macos
 
 			# check /usr/local/Homebrew
 
@@ -351,7 +366,13 @@ installDeps()
 
 			brew install $DEPS_MACOS
 
+			# Hook: after-install-deps-macos
+			includeExtensions after-install-deps-macos
+
 		elif [ $PLATFORM == "freebsd" ]; then
+
+			# Hook: before-install-deps-freebsd
+			includeExtensions before-install-deps-freebsd
 
 #			for p in "${!BINARY_PATHS_ARRAY[@]}" ; do
 #				if [ -f "${BINARY_PATHS_ARRAY[$p]}/git" ]; then
@@ -377,37 +398,37 @@ installDeps()
 				make BATCH=yes install clean
 			fi
 
-			if [ $ISSET_jpegoptim -eq 0 ]; then
+			if ! [ -z $ISSET_jpegoptim ] && [ $ISSET_jpegoptim -eq 0 ]; then
 				cd /usr/ports/graphics/jpegoptim/
 				make BATCH=yes install clean
 			fi
 
-			if [[ $ISSET_djpeg -eq 0 || $ISSET_cjpeg -eq 0 || $ISSET_jpegtran -eq 0 ]]; then
+			if ! [[ -z $ISSET_djpeg || -z $ISSET_cjpeg || -z $ISSET_jpegtran ]] && [[ $ISSET_djpeg -eq 0 || $ISSET_cjpeg -eq 0 || $ISSET_jpegtran -eq 0 ]]; then
 				cd /usr/ports/graphics/jpeg/
 				make BATCH=yes install clean
 			fi
 
-			if [ $ISSET_pngcrush -eq 0 ]; then
+			if ! [ -z $ISSET_pngcrush ] && [ $ISSET_pngcrush -eq 0 ]; then
 				cd /usr/ports/graphics/pngcrush/
 				make BATCH=yes install clean
 			fi
 
-			if [ $ISSET_optipng -eq 0 ]; then
+			if ! [ -z $ISSET_optipng ] && [ $ISSET_optipng -eq 0 ]; then
 				cd /usr/ports/graphics/optipng/
 				make BATCH=yes install clean
 			fi
 
-			if [ $ISSET_advpng -eq 0 ]; then
+			if ! [ -z $ISSET_advpng ] && [ $ISSET_advpng -eq 0 ]; then
 				cd /usr/ports/archivers/advancecomp/
 				make BATCH=yes install clean
 			fi
 
-			if [ $ISSET_gifsicle -eq 0 ]; then
+			if ! [ -z $ISSET_gifsicle ] && [ $ISSET_gifsicle -eq 0 ]; then
 				cd /usr/ports/graphics/gifsicle/
 				make BATCH=yes install clean
 			fi
 
-			if [ $ISSET_pngout -eq 0 ]; then
+			if ! [ -z $ISSET_pngout ] && [ $ISSET_pngout -eq 0 ]; then
 				cd ~
 				wget http://static.jonof.id.au/dl/kenutils/pngout-20150319-bsd.tar.gz
 				tar -xf pngout-20150319-bsd.tar.gz
@@ -419,6 +440,9 @@ installDeps()
 				fi
 				rm -rf pngout-20150319-bsd
 			fi
+
+			# Hook: after-install-deps-freebsd
+			includeExtensions after-install-deps-freebsd
 
 		fi
 
@@ -618,16 +642,37 @@ includeExtensions()
 	if ! [ -z "$1" ]; then
 		local EXTF_LIST=$(grep -lr "^#\ Hook:\ $1$" extensions | tr '\n' ' ' | sed 's/\ $//')
 		if ! [ -z "$EXTF_LIST" ]; then
-			local EXTF_ARR=($EXTF_LIST)
+			local EXTF_ARR=("$EXTF_LIST")
 			for EXTF in $EXTF_ARR; do
-				if [ $DEBUG -eq 1 ]; then
-					echo -n "including $(basename $EXTF)"
-				fi
+#				if [ $DEBUG -eq 1 ]; then
+#					echo -n "including $(basename $EXTF)"
+#				fi
 				. "$EXTF"
 			done
 		fi
 	fi
 }
+
+createLockFile()
+{
+	echo
+}
+
+removeLockFile()
+{
+	echo
+}
+
+#arrayContains () { 
+#	local list=$1[@]
+#	local elem=$2
+#	for i in "${!list}"; do
+#		if [ "$i" == "${elem}" ] ; then
+#			return 0
+#		fi
+#	done
+#	return 1
+#}
 
 usage()
 {
@@ -688,7 +733,9 @@ usage()
 	echo
 }
 
-# Define inner default vars. Don't change them!
+# Define default script vars
+TMP_PATH="/tmp"
+GIT_URL="https://github.com/zevilz/zImageOptimizer"
 DEBUG=0
 HELP=0
 SHOW_VERSION=0
@@ -699,9 +746,123 @@ PERIOD=0
 NEW_ONLY=0
 TIME_MARKER=""
 EXCLUDE_LIST=""
+ALL_FOUND=1
 PARAMS_NUM=$#
 SCRIPT_PATH="$( cd "$(dirname "$0")" ; pwd -P )"
 
+# Register image types
+declare -A IMG_TYPES_ARR
+IMG_TYPES_ARR[JPG]="JPG"
+IMG_TYPES_ARR[PNG]="PNG"
+IMG_TYPES_ARR[GIF]="GIF"
+
+# Hook: after-init-image-types
+includeExtensions after-init-image-types
+
+# Register tools
+declare -A TOOLS
+if ! [ -z "${IMG_TYPES_ARR[JPG]}" ]; then
+	TOOLS[JPG]="jpegoptim jpegtran djpeg cjpeg"
+fi
+if ! [ -z "${IMG_TYPES_ARR[PNG]}" ]; then
+	TOOLS[PNG]="pngcrush optipng pngout advpng"
+fi
+if ! [ -z "${IMG_TYPES_ARR[GIF]}" ]; then
+	TOOLS[GIF]="gifsicle"
+fi
+
+# Hook: after-init-tools
+includeExtensions after-init-tools
+
+# Generate tools array
+TOOLS_ARRAY=($(echo ${TOOLS[@]}))
+
+# Hook: after-init-tools-array
+includeExtensions after-init-tools-array
+
+# Register OS-based dependencies
+declare -A DEPS_DEBIAN_ARR
+declare -A DEPS_REDHAT_ARR
+declare -A DEPS_MACOS_ARR
+DEPS_DEBIAN="wget autoconf automake libtool make bc"
+DEPS_REDHAT="wget autoconf automake libtool make bc"
+DEPS_MACOS=""
+if ! [ -z "${TOOLS[JPG]}" ]; then
+	DEPS_DEBIAN_ARR[JPG]="jpegoptim libjpeg-progs"
+	DEPS_REDHAT_ARR[JPG]="jpegoptim libjpeg*"
+	DEPS_MACOS_ARR[JPG]="jpegoptim libjpeg"
+fi
+if ! [ -z "${TOOLS[PNG]}" ]; then
+	DEPS_DEBIAN_ARR[PNG]="pngcrush optipng advancecomp"
+	DEPS_REDHAT_ARR[PNG]="pngcrush optipng advancecomp"
+	DEPS_MACOS_ARR[PNG]="pngcrush optipng advancecomp jonof/kenutils/pngout"
+fi
+if ! [ -z "${TOOLS[GIF]}" ]; then
+	DEPS_DEBIAN_ARR[GIF]="gifsicle"
+	DEPS_REDHAT_ARR[GIF]="gifsicle"
+	DEPS_MACOS_ARR[GIF]="gifsicle"
+fi
+
+# Hook: after-init-deps
+includeExtensions after-init-deps
+
+# Generate OS-based dependencies
+for DEPS_DEBIAN_ITEM in "${DEPS_DEBIAN_ARR[@]}"; do
+	DEPS_DEBIAN="${DEPS_DEBIAN} ${DEPS_DEBIAN_ITEM}"
+done
+for DEPS_REDHAT_ITEM in "${DEPS_REDHAT_ARR[@]}"; do
+	DEPS_REDHAT="${DEPS_REDHAT} ${DEPS_REDHAT_ITEM}"
+done
+for DEPS_MACOS_ITEM in "${DEPS_MACOS_ARR[@]}"; do
+	DEPS_MACOS="${DEPS_MACOS} ${DEPS_MACOS_ITEM}"
+done
+
+# Register binary paths
+BINARY_PATHS="/bin /usr/bin /usr/local/bin"
+
+# Hook: after-init-binary-paths
+includeExtensions after-init-binary-paths
+
+# Generate binary paths array
+BINARY_PATHS=$(echo $BINARY_PATHS | sed 's/\/\ /\ /g' | sed 's/\/$/\ /')
+BINARY_PATHS_ARRAY=($BINARY_PATHS)
+
+# Hook: after-init-binary-paths-array
+includeExtensions after-init-binary-paths-array
+
+# Register time marker vars
+TIME_MARKER_PATH=""
+TIME_MARKER_NAME=".timeMarker"
+
+# Register min versions of Linux distros. Must be integer.
+MIN_VERSION_DEBIAN=7
+MIN_VERSION_UBUNTU=14
+MIN_VERSION_FEDORA=24
+MIN_VERSION_RHEL=6
+MIN_VERSION_CENTOS=6
+
+# Register min version MacOS (second digit; ex. 10.12.2 == 12).
+MIN_VERSION_MACOS=10
+
+# Register spacese separated supported versions of FreeBSD.
+SUPPORTED_VERSIONS_FREEBSD="10.3 10.4 11.1"
+
+# Define CRON and direct using styling
+if [ "Z$(ps o comm="" -p $(ps o ppid="" -p $$))" == "Zcron" -o \
+     "Z$(ps o comm="" -p $(ps o ppid="" -p $(ps o ppid="" -p $$)))" == "Zcron" ]; then
+	SETCOLOR_SUCCESS=
+	SETCOLOR_FAILURE=
+	SETCOLOR_NORMAL=
+else
+	SETCOLOR_SUCCESS="echo -en \\033[1;32m"
+	SETCOLOR_FAILURE="echo -en \\033[1;31m"
+	SETCOLOR_NORMAL="echo -en \\033[0;39m"
+fi
+
+# Hook: after-init-vars
+includeExtensions after-init-vars
+
+# Parse options
 while [ 1 ] ; do
 	if [ "${1#--path=}" != "$1" ] ; then
 		DIR_PATH="${1#--path=}"
@@ -760,17 +921,22 @@ while [ 1 ] ; do
 	shift
 done
 
+includeExtensions after-parse-params
+
+# Show help
 if [[ $HELP -eq 1 || $PARAMS_NUM -eq 0 ]]; then
 	usage
 	exit 0
 fi
 
+# Show version
 if [ $SHOW_VERSION -eq 1 ]; then
 	CUR_VERSION=$(grep 'Version:\ ' $0 | cut -d ' ' -f3)
 	echo $CUR_VERSION
 	exit 0
 fi
 
+# Checking input data
 if [ $CHECK_ONLY -eq 0 ]; then
 
 	DIR_PATH=$(echo "$DIR_PATH" | sed 's/\/$//')
@@ -863,12 +1029,9 @@ if [ $CHECK_ONLY -eq 0 ]; then
 
 fi
 
-BINARY_PATHS=$(echo $BINARY_PATHS | sed 's/\/\ /\ /g' | sed 's/\/$/\ /')
-BINARY_PATHS_ARRAY=($BINARY_PATHS)
-TOOLS_ARRAY=($TOOLS)
-ALL_FOUND=1
-
 echo
+
+# Checking tools
 echo -n "Checking tools"
 if [ $DEBUG -eq 1 ]; then
 	echo -n " in $BINARY_PATHS"
@@ -908,6 +1071,7 @@ done
 
 echo
 
+# Dialogs after checking tools
 if [ $ALL_FOUND -eq 1 ]; then
 	echo "All tools found"
 	echo
@@ -973,6 +1137,7 @@ else
 	fi
 fi
 
+# Find images
 IMAGES=$(\
 find "$DIR_PATH" $FIND_INCLUDE \( \
 -name '*.jpg' -or \
@@ -985,6 +1150,7 @@ find "$DIR_PATH" $FIND_INCLUDE \( \
 -name '*.PNG' \
 \) | findExclude)
 
+# Num of images
 IMAGES_TOTAL=$(\
 find "$DIR_PATH" $FIND_INCLUDE \( \
 -name '*.jpg' -or \
@@ -997,23 +1163,29 @@ find "$DIR_PATH" $FIND_INCLUDE \( \
 -name '*.PNG' \
 \) | findExclude | wc -l)
 
+# Preoptimize vars
 IMAGES_OPTIMIZED=0
 IMAGES_CURRENT=0
 START_TIME=$(date +%s)
 
+# If images found
 if ! [ -z "$IMAGES" ]; then
 
 	echo "Optimizing..."
 
+	# Init stat vars
 	INPUT=0
 	OUTPUT=0
 	SAVED_SIZE=0
 
+	# Main optimize loop
 	echo "$IMAGES" | ( while read IMAGE ; do
 
+		# Define additional vars for using hooks
 		OPTIMIZE=1
 		COMPARE_SIZE_AFTER=1
 
+		# Process counter
 		if [ $LESS -eq 0 ]; then
 #			if [ $SHOW_PROGRESS -eq 1 ]; then
 #				if [ $PROGRESS_MEASURE == "percent" ]; then
@@ -1033,13 +1205,16 @@ if ! [ -z "$IMAGES" ]; then
 			echo -n "$IMAGE"
 			echo -n '... '
 		fi
+
+		# Sizes before optimizing
 		SIZE_BEFORE=$(wc -c "$IMAGE" | awk '{print $1}')
 		SIZE_BEFORE_SCALED=$(echo "scale=1; $SIZE_BEFORE/1024" | bc | sed 's/^\./0./')
 		INPUT=$(echo "$INPUT+$SIZE_BEFORE" | bc)
 
+		# Get image extension
 		EXT=${IMAGE##*.}
 
-		# save permissions
+		# Save permissions
 		if [[ "$OSTYPE" == "linux-gnu" ]]; then
 			CUR_OWNER=$(stat -c "%U:%G" "$IMAGE")
 			CUR_PERMS=$(stat -c "%a" "$IMAGE")
@@ -1049,20 +1224,21 @@ if ! [ -z "$IMAGES" ]; then
 			CUR_PERMS=$(stat -f "%Lp" "$IMAGE")
 		fi
 
-		# save original file
+		# Save original file
 		cp -f "$IMAGE" "$TMP_PATH/$(basename "$IMAGE").bkp"
 
+		# JPEG
 		if [[ $EXT == "jpg" || $EXT == "jpeg" || $EXT == "JPG" || $EXT == "JPEG" ]]; then
 
 			includeExtensions optim-jpg-before
 
 			if [ $OPTIMIZE -eq 1 ]; then
 
-				if [ $ISSET_jpegoptim -eq 1 ]; then
+				if [[ $ISSET_jpegoptim -eq 1 ]]; then
 					optimJpegoptim "$IMAGE"
 				fi
 
-				if [ $ISSET_jpegtran -eq 1 ]; then
+				if [[ $ISSET_jpegtran -eq 1 ]]; then
 					optimJpegtran "$IMAGE"
 				fi
 
@@ -1074,29 +1250,30 @@ if ! [ -z "$IMAGES" ]; then
 
 			includeExtensions optim-jpg-after
 
+		# PNG
 		elif [[ $EXT == "png" || $EXT == "PNG" ]]; then
 
 			includeExtensions optim-png-before
 
 			if [ $OPTIMIZE -eq 1 ]; then
 
-		#		if [ $ISSET_convert -eq 1 ]; then
+		#		if [[ $ISSET_convert -eq 1 ]]; then
 		#			optimConvert "$IMAGE"
 		#		fi
 
-				if [ $ISSET_pngcrush -eq 1 ]; then
+				if [[ $ISSET_pngcrush -eq 1 ]]; then
 					optimPngcrush "$IMAGE"
 				fi
 
-				if [ $ISSET_optipng -eq 1 ]; then
+				if [[ $ISSET_optipng -eq 1 ]]; then
 					optimOptipng "$IMAGE"
 				fi
 
-				if [ $ISSET_pngout -eq 1 ]; then
+				if [[ $ISSET_pngout -eq 1 ]]; then
 					optimPngout "$IMAGE"
 				fi
 
-				if [ $ISSET_advpng -eq 1 ]; then
+				if [[ $ISSET_advpng -eq 1 ]]; then
 					optimAdvpng "$IMAGE"
 				fi
 
@@ -1104,13 +1281,14 @@ if ! [ -z "$IMAGES" ]; then
 
 			includeExtensions optim-png-after
 
+		# GIF
 		elif [[ $EXT == "gif" || $EXT == "GIF" ]]; then
 
 			includeExtensions optim-gif-before
 
 			if [ $OPTIMIZE -eq 1 ]; then
 
-				if [ $ISSET_gifsicle -eq 1 ]; then
+				if [[ $ISSET_gifsicle -eq 1 ]]; then
 					optimGifsicle "$IMAGE"
 				fi
 
@@ -1120,12 +1298,13 @@ if ! [ -z "$IMAGES" ]; then
 
 		fi
 
+		# Sizes after
 		SIZE_AFTER=$(wc -c "$IMAGE" | awk '{print $1}')
 		SIZE_AFTER_SCALED=$(echo "scale=1; $SIZE_AFTER/1024" | bc | sed 's/^\./0./')
 
+		# Compare original and optimized filesize
 		if [ $COMPARE_SIZE_AFTER -eq 1 ]; then
 
-			# compare original and optimized filesize
 			if [ $SIZE_BEFORE -le $SIZE_AFTER ]; then
 				OUTPUT=$(echo "$OUTPUT+$SIZE_BEFORE" | bc)
 				cp -f "$TMP_PATH/$(basename "$IMAGE").bkp" "$IMAGE"
@@ -1135,13 +1314,14 @@ if ! [ -z "$IMAGES" ]; then
 
 		fi
 
-		# restore permissions
+		# Restore permissions
 		chown $CUR_OWNER "$IMAGE"
 		chmod $CUR_PERMS "$IMAGE"
 
-		# remove original file
+		# Remove original file
 		rm "$TMP_PATH/$(basename "$IMAGE").bkp"
 
+		# Optimize results and sizes
 		if [ $SIZE_BEFORE -le $SIZE_AFTER ]; then
 			if [ $LESS -eq 0 ]; then
 				$SETCOLOR_FAILURE
@@ -1167,6 +1347,7 @@ if ! [ -z "$IMAGES" ]; then
 
 	done
 
+	# Total info
 	echo
 	echo -n "Input: "
 	readableSize $INPUT
@@ -1190,6 +1371,8 @@ if ! [ -z "$IMAGES" ]; then
 	echo -n "Total optimizing time: "
 	readableTime $TOTAL_TIME
 	)
+
+	# Update time marker
 	updateTimeMarker
 
 else
